Fix documentation for RemoveContactFromList
Hello, I believe there was a bug in the documentation. It looks like a copy/paste from another function that was not updated (probably from CreateContacts). 
I've updated the "ids" property description to be similar to the "emails" param.
Please double-check that the edits I made are accurate and reflect expected API behavior.